trading/requirements.txt
pandas
numpy
scikit-learn
pyyaml
python-dotenv
streamlit
plotly
fyers-apiv3
